Implemented various fixes
Changed the method of adding classifications together at Fuzzy's request.
Corrected the order that Progression and Progression + Skip Balancing were checked so that Progression + Skip Balancing takes priority again.
Modified a check so that it now looks for if an item has the Progression flag, rather than being exactly Progression or Progression + Skip Balancing.
Added a case for removing Useful + Trap items if there aren't enough locations for them in the pool.